feat: remove redundant type cast

+++ src/main/java/com/onixbyte/captcha/text/impl/DefaultWordRenderer.java
@@ -110,7 +110,7 @@ public class DefaultWordRenderer implements WordRenderer {
                     wordChars[i]
             };
             g2D.drawChars(charToDraw, 0, charToDraw.length, startPosX, startPosY);
-            startPosX = startPosX + (int) charWidths[i] + charSpace;
+            startPosX = startPosX + charWidths[i] + charSpace;
         }
 
         return image;
